id column datatype mismatch handled

+++ quiz-app-backend/src/main/java/sm/quiz/entities/dto/AnswerResult.java
@@ -10,5 +10,5 @@ public class AnswerResult {
 	private Long questionId;
     private boolean correct;
     private Integer score;
-    private Set<Long> correctOptionIds;
+    private Set<Integer> correctOptionIds;
 }
