+++ quiz-app-backend/src/main/java/sm/quiz/entities/dto/AnswerSubmission.java
@@ -8,6 +8,6 @@ import lombok.Data;
 public class AnswerSubmission {
 
 	private Long questionId;
-    private Set<Long> selectedOptionIds;
+    private Set<Integer> selectedOptionIds;
 
 }
